requirements2.txt
tqdm
streamlit
pandas
numpy
openpyxl
pyreadstat
plotly
openai
ollama
pdf2image
pillow
regex
PyPDF2
pdfplumber
python-docx
docx2pdf
pytesseract
requests
lxml
scikit-learn
